+++ src/postgres/teams/invitation.rs
@@ -32,12 +32,12 @@ struct InvitationRecord {
 impl From<InvitationRecord> for TeamInvitation {
     fn from(row: InvitationRecord) -> Self {
         TeamInvitation {
-            id: row.id as u64,
-            team_id: row.team_id as u64,
+            id: row.id,
+            team_id: row.team_id,
             email: row.email,
             role: row.role,
             token_hash: row.token_hash,
-            invited_by: row.invited_by as u64,
+            invited_by: row.invited_by,
             expires_at: row.expires_at,
             accepted_at: row.accepted_at,
             created_at: row.created_at,
@@ -56,11 +56,11 @@ impl TeamInvitationRepository for PostgresTeamInvitationRepository {
             RETURNING id, team_id, email, role, token_hash, invited_by, expires_at, accepted_at, created_at
             ",
         )
-        .bind(data.team_id as i64)
+        .bind(data.team_id)
         .bind(&data.email)
         .bind(&data.role)
         .bind(&data.token_hash)
-        .bind(data.invited_by as i64)
+        .bind(data.invited_by)
         .bind(data.expires_at)
         .fetch_one(&self.pool)
         .await
@@ -73,11 +73,11 @@ impl TeamInvitationRepository for PostgresTeamInvitationRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn find_by_id(&self, id: u64) -> Result<Option<TeamInvitation>, AuthError> {
+    async fn find_by_id(&self, id: i64) -> Result<Option<TeamInvitation>, AuthError> {
         let row: Option<InvitationRecord> = sqlx::query_as(
             "SELECT id, team_id, email, role, token_hash, invited_by, expires_at, accepted_at, created_at FROM team_invitations WHERE id = $1",
         )
-        .bind(id as i64)
+        .bind(id)
         .fetch_optional(&self.pool)
         .await
         .map_err(|e| {
@@ -108,7 +108,7 @@ impl TeamInvitationRepository for PostgresTeamInvitationRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn find_pending_by_team(&self, team_id: u64) -> Result<Vec<TeamInvitation>, AuthError> {
+    async fn find_pending_by_team(&self, team_id: i64) -> Result<Vec<TeamInvitation>, AuthError> {
         let rows: Vec<InvitationRecord> = sqlx::query_as(
             r"
             SELECT id, team_id, email, role, token_hash, invited_by, expires_at, accepted_at, created_at
@@ -117,7 +117,7 @@ impl TeamInvitationRepository for PostgresTeamInvitationRepository {
             ORDER BY created_at DESC
             ",
         )
-        .bind(team_id as i64)
+        .bind(team_id)
         .fetch_all(&self.pool)
         .await
         .map_err(|e| {
@@ -150,7 +150,7 @@ impl TeamInvitationRepository for PostgresTeamInvitationRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn mark_accepted(&self, id: u64) -> Result<TeamInvitation, AuthError> {
+    async fn mark_accepted(&self, id: i64) -> Result<TeamInvitation, AuthError> {
         let row: InvitationRecord = sqlx::query_as(
             r"
             UPDATE team_invitations SET accepted_at = NOW()
@@ -158,7 +158,7 @@ impl TeamInvitationRepository for PostgresTeamInvitationRepository {
             RETURNING id, team_id, email, role, token_hash, invited_by, expires_at, accepted_at, created_at
             ",
         )
-        .bind(id as i64)
+        .bind(id)
         .fetch_one(&self.pool)
         .await
         .map_err(|e| match e {
@@ -173,9 +173,9 @@ impl TeamInvitationRepository for PostgresTeamInvitationRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn delete(&self, id: u64) -> Result<(), AuthError> {
+    async fn delete(&self, id: i64) -> Result<(), AuthError> {
         sqlx::query("DELETE FROM team_invitations WHERE id = $1")
-            .bind(id as i64)
+            .bind(id)
             .execute(&self.pool)
             .await
             .map_err(|e| {
@@ -187,7 +187,7 @@ impl TeamInvitationRepository for PostgresTeamInvitationRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn delete_expired(&self) -> Result<u64, AuthError> {
+    async fn delete_expired(&self) -> Result<i64, AuthError> {
         let result = sqlx::query("DELETE FROM team_invitations WHERE expires_at < NOW()")
             .execute(&self.pool)
             .await
@@ -196,6 +196,6 @@ impl TeamInvitationRepository for PostgresTeamInvitationRepository {
                 AuthError::DatabaseError(e.to_string())
             })?;
 
-        Ok(result.rows_affected())
+        Ok(result.rows_affected() as i64)
     }
 }
