+++ src/postgres/teams/membership.rs
@@ -29,9 +29,9 @@ struct MembershipRecord {
 impl From<MembershipRecord> for TeamMembership {
     fn from(row: MembershipRecord) -> Self {
         TeamMembership {
-            id: row.id as u64,
-            team_id: row.team_id as u64,
-            user_id: row.user_id as u64,
+            id: row.id,
+            team_id: row.team_id,
+            user_id: row.user_id,
             role: row.role,
             created_at: row.created_at,
             updated_at: row.updated_at,
@@ -50,8 +50,8 @@ impl TeamMembershipRepository for PostgresTeamMembershipRepository {
             RETURNING id, team_id, user_id, role, created_at, updated_at
             ",
         )
-        .bind(data.team_id as i64)
-        .bind(data.user_id as i64)
+        .bind(data.team_id)
+        .bind(data.user_id)
         .bind(&data.role)
         .fetch_one(&self.pool)
         .await
@@ -64,11 +64,11 @@ impl TeamMembershipRepository for PostgresTeamMembershipRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn find_by_id(&self, id: u64) -> Result<Option<TeamMembership>, AuthError> {
+    async fn find_by_id(&self, id: i64) -> Result<Option<TeamMembership>, AuthError> {
         let row: Option<MembershipRecord> = sqlx::query_as(
             "SELECT id, team_id, user_id, role, created_at, updated_at FROM team_memberships WHERE id = $1",
         )
-        .bind(id as i64)
+        .bind(id)
         .fetch_optional(&self.pool)
         .await
         .map_err(|e| {
@@ -82,14 +82,14 @@ impl TeamMembershipRepository for PostgresTeamMembershipRepository {
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
     async fn find_by_team_and_user(
         &self,
-        team_id: u64,
-        user_id: u64,
+        team_id: i64,
+        user_id: i64,
     ) -> Result<Option<TeamMembership>, AuthError> {
         let row: Option<MembershipRecord> = sqlx::query_as(
             "SELECT id, team_id, user_id, role, created_at, updated_at FROM team_memberships WHERE team_id = $1 AND user_id = $2",
         )
-        .bind(team_id as i64)
-        .bind(user_id as i64)
+        .bind(team_id)
+        .bind(user_id)
         .fetch_optional(&self.pool)
         .await
         .map_err(|e| {
@@ -101,11 +101,11 @@ impl TeamMembershipRepository for PostgresTeamMembershipRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn find_by_team(&self, team_id: u64) -> Result<Vec<TeamMembership>, AuthError> {
+    async fn find_by_team(&self, team_id: i64) -> Result<Vec<TeamMembership>, AuthError> {
         let rows: Vec<MembershipRecord> = sqlx::query_as(
             "SELECT id, team_id, user_id, role, created_at, updated_at FROM team_memberships WHERE team_id = $1 ORDER BY created_at ASC",
         )
-        .bind(team_id as i64)
+        .bind(team_id)
         .fetch_all(&self.pool)
         .await
         .map_err(|e| {
@@ -117,11 +117,11 @@ impl TeamMembershipRepository for PostgresTeamMembershipRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn find_by_user(&self, user_id: u64) -> Result<Vec<TeamMembership>, AuthError> {
+    async fn find_by_user(&self, user_id: i64) -> Result<Vec<TeamMembership>, AuthError> {
         let rows: Vec<MembershipRecord> = sqlx::query_as(
             "SELECT id, team_id, user_id, role, created_at, updated_at FROM team_memberships WHERE user_id = $1 ORDER BY created_at ASC",
         )
-        .bind(user_id as i64)
+        .bind(user_id)
         .fetch_all(&self.pool)
         .await
         .map_err(|e| {
@@ -133,7 +133,7 @@ impl TeamMembershipRepository for PostgresTeamMembershipRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn update_role(&self, id: u64, role: &str) -> Result<TeamMembership, AuthError> {
+    async fn update_role(&self, id: i64, role: &str) -> Result<TeamMembership, AuthError> {
         let row: MembershipRecord = sqlx::query_as(
             r"
             UPDATE team_memberships SET role = $1, updated_at = NOW()
@@ -142,7 +142,7 @@ impl TeamMembershipRepository for PostgresTeamMembershipRepository {
             ",
         )
         .bind(role)
-        .bind(id as i64)
+        .bind(id)
         .fetch_one(&self.pool)
         .await
         .map_err(|e| match e {
@@ -157,9 +157,9 @@ impl TeamMembershipRepository for PostgresTeamMembershipRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn delete(&self, id: u64) -> Result<(), AuthError> {
+    async fn delete(&self, id: i64) -> Result<(), AuthError> {
         sqlx::query("DELETE FROM team_memberships WHERE id = $1")
-            .bind(id as i64)
+            .bind(id)
             .execute(&self.pool)
             .await
             .map_err(|e| {
@@ -171,10 +171,10 @@ impl TeamMembershipRepository for PostgresTeamMembershipRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn delete_by_team_and_user(&self, team_id: u64, user_id: u64) -> Result<(), AuthError> {
+    async fn delete_by_team_and_user(&self, team_id: i64, user_id: i64) -> Result<(), AuthError> {
         sqlx::query("DELETE FROM team_memberships WHERE team_id = $1 AND user_id = $2")
-            .bind(team_id as i64)
-            .bind(user_id as i64)
+            .bind(team_id)
+            .bind(user_id)
             .execute(&self.pool)
             .await
             .map_err(|e| {
